feat: add solution to the modules task
